move buddy list population from OService.ClientOnline to Feedbag.Use
This is the first step towards supporting both client-side and
server-side buddy list management. Prior to this commit, the buddy list
is populated using the feedbag mechanism in OService.ClientOnline. This
logic is redundant and can cause confusion for older AIM clients that
use client-side buddy list management. Instead, we now populate the
buddy list using feedbag only when a feedbag-enabled client tells to via
the Feedbag.Use SNAC.