support client-side buddy lists for AIM 4.1
AIM versions 4.1 and older use client-side buddy list management and
versions newer than 4.1 use feedbag/server-side buddy list management.

The basic flow of this client-side buddy list implementation:
At signin, the AIM client dumps the buddy list content via SNAC
0x03-0x04 AddBuddies. The server adds each User->Buddy association to
an in-memory adjacency list. This list is used to retrieve a user's
buddies as well as look up all the users who have added a user as a
buddy.

When the server sends BuddyArrived and BuddyDeparted events, it consults
the adjacency list (in addition to the feedbag) for a list of users.