support md5 password hash for AIM 3.5-4.7
There are two kinds of password hashing for BUCP auth.

The first type is what we'll call the the "weak" password hash.
It's new to this code base and unlocks support for AIM 3.5-4.7

  md5(auth key + password + "AOL Instant Messenger (SM)")

The second type is what we'll call the "strong" password hash, and
already existed before this commit:

  md5(auth key + md5(password) + "AOL Instant Messenger (SM)")

Now when a new account is created, both types of hashes are generated
for each password. At login time, the auth service checks both hashes
and succeeds if the hash provided by the client matches one of these
hashes.

Unfortunately, this requires server operators to reset the passwords for
all users in order to generate the weak password hash, as it's not
possible to derive the weak password hash from the existing strong
password hash.