issue #14 - implement SNAC 09,03 PD RightsQuery
This commit implements the first handler in the Permit/Deny foodgroup,
RightsQuery, which allows Pidgin 2.12.0 login flow to successfully
complete.

This work reveals that the Permit/Deny food group needs to be
fully implemented in order to support user blocking in non-feedbag
AIM clients.

For now, we'll get PD out of the list of available foodgroups at login
until this is sorted out.