resolve confusion over which executable to run
Some users reported that RAS was not picking up changes to their
settings files. This occurred when they would start the application
executable rather than the start script, which loads the environment
file before starting the application. When this happened, the
application loaded setting defaults.

To fix this issue, this commit does two things to ensure that users run
the application through the script:
- Move the binary into a bin/ subfolder
- Remove default settings entirely, so that the application exits with
  an error if no settings are applied.