fix incorrect user idle duration in UserInfo
WinAIM clients report idle duration that exceeds online time by hours.
Pidgin displays all users as idle when they are not idle.

Fix here is to set the correct duration for user info TLV 0x04 in
minutes, not seconds.

In addition, we should not send TLV 0x04 if idle duration is 0. Pidgin
incorrectly interprets the presence of TLV 0x04 as the user being idle,
regardless if the value is 0.