converted buddy arrival/departure funcs to methods
This commit converts the following functions into methods on the Buddy
service:

- broadcastArrival
- broadcastDeparture
- unicastArrival
- unicastDeparture

These functions required extensive mocking of the MessageRelayer,
FeedbagManager, and LegacyBuddyListManager for each call.

Now that these functions are methods, it's easier to mock them rather
than mock all of their dependencies. This reduces a lot of effort needed
to write unit tests for functions that send buddy arrival/departure
notifications.