remove bart flags check from RetrieveItem.RetrieveItem
Previously, RetrieveItem would return an error if it received an item
request with any flags set. The original thinking was that this method
should only receive requests for "known" BART item types.

Recently, a bug surfaced on Chivanet where a client, who had set a buddy
icon via iChat with flag 0x01 ("custom"), would force its buddies to
disconnect every time it signed on, because RetrieveItem did not accept
the flags.

It's not clear why RetrieveItem would take any flag into consideration.
This commit removes the check for now. Maybe it will become apparent in
the future why flags are passed to the RetrieveItem request.