Send OServiceUserInfoUpdate when formatting screenname
This is needed to inform the client that their user info has changed. For example,
when viewing a profile with '%n' it was incorrectly still showing the old format.

The buddy list header still shows the old format on the 5.1.3036 client but may be a
client bug. It works correctly on a 5.9 client.