wire: add optional struct tag
This commit adds a new "optional" struct tag.

struct {
        Val0     uint16
        Optional *struct {
                Val1 uint16
        } `oscar:"optional"`
}

The marshaler skips marshaling pointer struct fields marked optional if
the pointer is nil. No bytes are written to the output byte array.

The unmarshaler skips unmarshaling pointer struct fields marked
optional if the pointer is nil. It leaves the pointer nil and does not
throw an EOF error, which is the default behavior for non-optional
fields.

The optional field must be the -last- field in the struct, otherwise the
marshaler/unmarshaler throws an error.

In addition, this commit aligns the struct tag format more closely with
convention. All the marshaling options now belong under the "oscar" key.