load env file at application startup
Replace the run scripts run.sh and run.cmd that load settings from the
config files settings.env and settings.bat with a scheme the loads a
single env file from the application at startup for all platforms.

The impetus for this was that there was no way to make starting RAS
one-click under Windows using a script that could be codesigned. The
existing run.cmd file could not be codesigned, so the user was always
confronted with running an application from an "unknown" developer. The
next best option was to rewrite run.cmd as a powershell script, which
can be codesigned. This was not a great option because you can not start
a powershell script by default by double-clicking it. You have to
instead right-click the script and click "run in powershell".

In order to streamline the user experience, I wanted to find a way to
start a codesigned application with a double-click. The best option,
implemented in this commit, is to load the settings file at startup
within the signed golang application.

As a result, we have less configuration to maintain, and Windows users
will be able to more easily start the application, while still retaining
the ability to configure the application via environment variables.