force shutdown if connections don't close
There is a hard to reproduce bug in the shutdown logic where the
application hangs on SIGINT shutdown. I think this might occur when the
server blocks on writing to a dead TCP connection. I will explore the
root cause at a later date. For now, I'm adding a forced shutdown that
occurs 5 seconds after SIGINT is processed.