pd: set permit/deny mode on empty list
I ran into a situation where the permit-some and deny-some modes
are not set when the user list is empty. This behavior surfaced while
testing the TIK client. This does not occur in the regular OSCAR AIM
client because it sets "allow all" when the "deny some" list is empty
and "deny all" when the "permit some" list is empty.